\name{matter-utils}
\docType{data}

\alias{matter-utils}

\alias{data:matter_ex}
\alias{matter_ex}
\alias{data:matter_sim}
\alias{matter_sim}
\alias{data:matter_msi}
\alias{matter_msi}

\alias{bigglm.out}
\alias{prcomp.out}
\alias{lm.prof}
\alias{pca.prof}
\alias{msi.prof}

\alias{sizeof}
\alias{make_datamode}
\alias{convert_datamode}
\alias{widest_datamode}
\alias{vm_used}

\title{Internal Utilities for ``matter'' Package}

\description{
  Low-level utility functions, classes, and data defined in the \pkg{matter} package. They are not intended to be used directly.
}

\keyword{utilities}
\keyword{classes}
\keyword{methods}
\keyword{datasets}
