\name{matter_ex-data}
\docType{data}

\alias{data:matter_ex}
\alias{matter_ex}

\alias{bigglm.out}
\alias{prcomp.out}

\title{Examples for ``matter'' package}

\description{
	Example data for the ``matter'' package for use in vignettes.
}

\usage{
data(matter_ex)
}

\value{
	None. Loads data objects required to build vignettes.
}

\keyword{datasets}
