\name{matter-utils}
\docType{data}

\alias{matter-utils}

\alias{data:matter_ex}
\alias{matter_ex}
\alias{data:matter_sim}
\alias{matter_sim}
\alias{data:matter_msi}
\alias{matter_msi}

\alias{bigglm.out}
\alias{prcomp.out}
\alias{lm.prof}
\alias{pca.prof}
\alias{msi.prof}

\alias{describe_for_display}
\alias{describe_for_display,ANY-method}
\alias{describe_for_display,atoms-method}
\alias{describe_for_display,drle-method}
\alias{describe_for_display,matter_vec-method}
\alias{describe_for_display,matter_arr-method}
\alias{describe_for_display,matter_mat-method}
\alias{describe_for_display,matter_list-method}
\alias{describe_for_display,matter_str-method}
\alias{describe_for_display,matter_fc-method}
\alias{describe_for_display,matter_df-method}
\alias{describe_for_display,rep_vt-method}
\alias{describe_for_display,sparse_mat-method}
\alias{describe_for_display,virtual_mat-method}
\alias{describe_for_display,virtual_tbl-method}

\alias{preview_for_display}
\alias{preview_for_display,ANY-method}
\alias{preview_for_display,atoms-method}
\alias{preview_for_display,drle-method}
\alias{preview_for_display,matter_vec-method}
\alias{preview_for_display,matter_arr-method}
\alias{preview_for_display,matter_mat-method}
\alias{preview_for_display,matter_list-method}
\alias{preview_for_display,matter_str-method}
\alias{preview_for_display,matter_fc-method}
\alias{preview_for_display,matter_df-method}
\alias{preview_for_display,rep_vt-method}
\alias{preview_for_display,sparse_mat-method}
\alias{preview_for_display,virtual_mat-method}
\alias{preview_for_display,virtual_tbl-method}

\alias{as.altrep}
\alias{as.altrep,ANY-method}
\alias{as.altrep,data.frame-method}
\alias{as.altrep,list-method}
\alias{as.altrep,matter_vec-method}
\alias{as.altrep,matter_arr-method}
\alias{as.altrep,matter_mat-method}
\alias{as.altrep,matter_list-method}
\alias{as.altrep,matter_str-method}
\alias{as.altrep,matter_fc-method}
\alias{as.altrep,virtual_df-method}
\alias{as.altrep,matter_df-method}

\alias{make_filemode}
\alias{common_filemode}

\alias{sizeof}
\alias{make_datamode}
\alias{convert_datamode}
\alias{widest_datamode}
\alias{vm_used}

\title{Internal Utilities for ``matter'' Package}

\description{
  Low-level utility functions, classes, and data defined in the \pkg{matter} package. They are not intended to be used directly.
}

\keyword{utilities}
\keyword{internal}
