\name{matter-utils}
\docType{data}

\alias{matter-utils}

\alias{describe_for_display}
\alias{describe_for_display,ANY-method}
\alias{describe_for_display,atoms-method}
\alias{describe_for_display,drle-method}
\alias{describe_for_display,drle_fct-method}
\alias{describe_for_display,matter_arr-method}
\alias{describe_for_display,matter_mat-method}
\alias{describe_for_display,matter_vec-method}
\alias{describe_for_display,matter_list-method}
\alias{describe_for_display,matter_str-method}
\alias{describe_for_display,matter_fct-method}
\alias{describe_for_display,sparse_vec-method}
\alias{describe_for_display,sparse_mat-method}

\alias{preview_for_display}
\alias{preview_for_display,ANY-method}
\alias{preview_for_display,atoms-method}
\alias{preview_for_display,drle-method}
\alias{preview_for_display,matter_arr-method}
\alias{preview_for_display,matter_mat-method}
\alias{preview_for_display,matter_vec-method}
\alias{preview_for_display,matter_list-method}
\alias{preview_for_display,matter_str-method}
\alias{preview_for_display,matter_fct-method}
\alias{preview_for_display,sparse_vec-method}
\alias{preview_for_display,sparse_mat-method}

\alias{as.altrep}
\alias{as.altrep,matter_vec-method}
\alias{as.altrep,matter_arr-method}
\alias{as.altrep,matter_mat-method}
\alias{as.altrep,matter_list-method}
\alias{as.altrep,matter_str-method}
\alias{as.altrep,matter_fct-method}

\alias{combine,ANY,ANY-method}

\alias{keys}
\alias{keys<-}
\alias{keys,ANY-method}
\alias{keys<-,ANY-method}

\alias{array_ind}
\alias{linear_ind}

\alias{update_attr}
\alias{copy_env}

\alias{size_bytes}

\alias{roll}
\alias{seq_rel}
\alias{shingles}
\alias{pinv}
\alias{avg}

\alias{parse_formula}
\alias{eval_exprs}
\alias{eval_at}
\alias{varquote}

\alias{panel_grid}
\alias{panel_dim}
\alias{panel_get}
\alias{panel_set}
\alias{panel_next}
\alias{panel_prev}

\alias{qrank}
\alias{qorder}
\alias{qselect}
\alias{qmedian}
\alias{qmad}

\alias{read_atom}
\alias{write_atom}

\alias{read_atoms}
\alias{write_atoms}

\alias{sizeof}

\alias{type,vector-method}
\alias{type,array-method}

\alias{shm_used}
\alias{shm_used,ANY-method}
\alias{shm_used,vector-method}
\alias{shm_used,array-method}
\alias{shm_used,atoms-method}
\alias{shm_used,matter_-method}
\alias{shm_used,sparse_arr-method}

\alias{vm_used}
\alias{vm_used,ANY-method}
\alias{vm_used,vector-method}
\alias{vm_used,array-method}
\alias{vm_used,atoms-method}
\alias{vm_used,matter_-method}
\alias{vm_used,sparse_arr-method}

\alias{vm_realized}
\alias{vm_realized,ANY-method}
\alias{vm_realized,matter_arr-method}
\alias{vm_realized,matter_list-method}
\alias{vm_realized,matter_str-method}
\alias{vm_realized,sparse_arr-method}

\alias{matter_shared_resource}
\alias{matter_shared_resource_list}
\alias{matter_shared_resource_pool}

\alias{matter_log}
\alias{matter_warn}
\alias{matter_error}

\title{Internal Utilities for ``matter'' Package}

\description{
  Low-level utility functions, classes, and data defined in the \pkg{matter} package that are exported for developer use only. They are not intended to be used directly.
}

\keyword{utilities}
\keyword{internal}
