\name{matter_ex-data}
\docType{data}

\alias{data:matter_ex}
\alias{matter_ex}
\alias{data:matter_sim}
\alias{matter_sim}
\alias{data:matter_msi}
\alias{matter_msi}

\alias{bigglm.out}
\alias{prcomp.out}
\alias{lm.prof}
\alias{pca.prof}
\alias{msi.prof}

\title{Example Data for ``matter'' Package}

\description{
	Example data for the ``matter'' package for use in vignettes.
}

\usage{
data(matter_ex)
data(matter_sim)
data(matter_msi)
}

\value{
	None. Loads objects required to build vignettes.
}

\keyword{datasets}
